+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - refactor: use a pre-fetched list of Quarto extensions for the QuickPick UI.
 - refactor: drop GitHub authentication requirement for the extension details.
+- refactor: change log level for cached extensions messages.
+- feat: add activation log message for Quarto Wizard.
 
 ## 0.8.1 (2025-02-20)